🖥️ ui(studio): enhance ChallengesPageView with pagination functionality and update useChallengesPage for improved state management
