+++ apps/studio/src/ui/challenging/Challenges/ChallengesPageView.tsx
@@ -19,11 +19,12 @@ import {
 import { Button } from '@/ui/shadcn/components/button'
 import { Badge } from '@/ui/shadcn/components/badge'
 import { ChallengesTableView } from '@/ui/global/widgets/components/ChallengesTable'
+import { Pagination } from '@/ui/global/widgets/components/Pagination'
 
 import { useChallengesPage } from './useChallengesPage'
 
 export const ChallengesPageView = () => {
-  const { challenges, isLoading, categories, filters } = useChallengesPage()
+  const { challenges, isLoading, categories, filters, pagination } = useChallengesPage()
 
   const toggleCategory = (categoryId: string) => {
     if (filters.selectedCategories.includes(categoryId)) {
@@ -101,9 +102,16 @@ export const ChallengesPageView = () => {
         </DropdownMenu>
       </div>
 
-      <div className='flex-1'>
-        <ChallengesTableView challenges={challenges} isLoading={isLoading} />
-      </div>
+      <ChallengesTableView challenges={challenges} isLoading={isLoading} />
+
+      <Pagination
+        page={pagination.page}
+        totalPages={pagination.totalPages}
+        totalItemsCount={pagination.totalItemsCount}
+        itemsPerPage={pagination.itemsPerPage}
+        onNextPage={pagination.handleNextPage}
+        onPrevPage={pagination.handlePrevPage}
+      />
     </div>
   )
 }
