+++ apps/studio/src/ui/challenging/Challenges/useChallengesPage.ts
@@ -27,27 +27,63 @@ export function useChallengesPage() {
     key: 'challenge-categories',
     fetcher: () => challengingService.fetchAllChallengeCategories(),
   })
+  const [page, setPage] = useState(1)
+  const itemsPerPage = 20
 
   const { data: challengesData, isLoading } = useCache({
     key: 'challenges-page-list',
-    dependencies: [debouncedSearch, difficulty, selectedCategories],
+    dependencies: [debouncedSearch, difficulty, selectedCategories, page],
     fetcher: async () =>
       await challengingService.fetchChallengesList({
-        page: OrdinalNumber.create(1),
-        itemsPerPage: OrdinalNumber.create(5),
+        page: OrdinalNumber.create(page),
+        itemsPerPage: OrdinalNumber.create(itemsPerPage),
         categoriesIds: IdsList.create(selectedCategories),
         completionStatus: ChallengeCompletionStatus.create('any'),
         difficulty: ChallengeDifficulty.create(difficulty),
         upvotesCountOrder: ListingOrder.create('any'),
         postingOrder: ListingOrder.create('desc'),
-        title: Text.create(debouncedSearch),
-        shouldIncludeStarChallenges: Logical.create(false),
         shouldIncludeOnlyAuthorChallenges: Logical.create(false),
+        shouldIncludeStarChallenges: Logical.create(false),
+        title: Text.create(debouncedSearch),
         userId: null,
       }),
   })
 
   const categories = categoriesData ?? []
+  const totalItemsCount = challengesData?.totalItemsCount ?? 0
+  const totalPages = Math.ceil(totalItemsCount / itemsPerPage)
+
+  const handleNextPage = () => {
+    if (page < totalPages) setPage((prev) => prev + 1)
+  }
+
+  const handlePrevPage = () => {
+    if (page > 1) setPage((prev) => prev - 1)
+  }
+
+  const handleSearchChange = (value: string) => {
+    setSearch(value)
+    setPage(1)
+  }
+
+  const handleDifficultyChange = (value: string) => {
+    setDifficulty(value)
+    setPage(1)
+  }
+
+  const handleCategoriesChange = (value: string[]) => {
+    setSelectedCategories(value)
+    setPage(1)
+  }
+
+  console.log({
+    page,
+    totalPages,
+    totalItemsCount,
+    itemsPerPage,
+    handleNextPage,
+    handlePrevPage,
+  })
 
   return {
     challenges: challengesData?.items ?? [],
@@ -55,11 +91,19 @@ export function useChallengesPage() {
     categories,
     filters: {
       search,
-      setSearch,
+      setSearch: handleSearchChange,
       difficulty,
-      setDifficulty,
+      setDifficulty: handleDifficultyChange,
       selectedCategories,
-      setSelectedCategories,
+      setSelectedCategories: handleCategoriesChange,
+    },
+    pagination: {
+      page,
+      totalPages,
+      totalItemsCount,
+      itemsPerPage,
+      handleNextPage,
+      handlePrevPage,
     },
   }
 }
