--- apps/studio/src/ui/global/widgets/pages/Challenges/index.tsx
@@ -1,3 +1,0 @@
-export const ChallengesPage = () => {
-  return <div>Challenges Page</div>
-}
